requirements.txt
streamlit-option-menu
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
colorama==0.4.6
cryptography==45.0.4
cycler==0.12.1
defusedxml==0.7.1
fonttools==4.58.4
greenlet==3.2.3
h11==0.16.0
huggingface-hub==0.33.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.9.0
mysql-connector-python==9.3.0
numpy==2.3.0
orjson==3.10.18
packaging==24.2
pandas==2.3.0
pdfplumber==0.11.7
Pillow==11.2.1
plotly==6.1.2
protobuf==6.31.1
psutil==7.0.0
pyarrow==20.0.0
pycparser==2.22
pydantic-core==2.33.2
pypdfium2==4.30.1
PyMuPDF==1.26.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
scikit-learn==1.7.0
scipy==1.15.3
six==1.17.0
SQLAlchemy==2.0.41
streamlit==1.45.1
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
transformers==4.52.4
typing-inspection==0.4.1
typing-extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
zstandard==0.23.0
vl-convert-python
fpdf2
